WIP for SPSA; Lots of code shuffling to clean it up
