+++ OpenBench/create_workload.py
@@ -1,0 +1,146 @@
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+#                                                                             #
+#   OpenBench is a chess engine testing framework authored by Andrew Grant.   #
+#   <https://github.com/AndyGrant/OpenBench>           <andrew@grantnet.us>   #
+#                                                                             #
+#   OpenBench is free software: you can redistribute it and/or modify         #
+#   it under the terms of the GNU General Public License as published by      #
+#   the Free Software Foundation, either version 3 of the License, or         #
+#   (at your option) any later version.                                       #
+#                                                                             #
+#   OpenBench is distributed in the hope that it will be useful,              #
+#   but WITHOUT ANY WARRANTY; without even the implied warranty of            #
+#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the             #
+#   GNU General Public License for more details.                              #
+#                                                                             #
+#   You should have received a copy of the GNU General Public License         #
+#   along with this program.  If not, see <http://www.gnu.org/licenses/>.     #
+#                                                                             #
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+import math
+
+import OpenBench.utils
+import OpenBench.verify_workload
+import OpenBench.views
+
+from OpenBench.models import *
+
+def create_workload(request, workload_type):
+
+    assert workload_type in [ 'TEST', 'TUNE' ]
+
+    if not request.user.is_authenticated:
+        return OpenBench.views.redirect(request, '/login/', error='Only enabled users can create tests')
+
+    if not Profile.objects.get(user=request.user).enabled:
+        return OpenBench.views.redirect(request, '/login/', error='Only enabled users can create tests')
+
+    if request.method == 'GET':
+        data     = { 'networks' : list(Network.objects.all().values()) }
+        template = 'create_test.html' if workload_type == 'TEST' else 'create_tune.html'
+        return OpenBench.views.render(request, template, data)
+
+    if workload_type == 'TEST':
+        workload, errors = create_new_test(request)
+
+    if workload_type == 'TUNE':
+        workload, errors = create_new_tune(request)
+
+    if errors != [] and errors != None:
+        url = '/newTest/' if workload_type == 'TEST' else '/newTune/'
+        return OpenBench.views.redirect(request, url, error='\n'.join(errors))
+
+    if warning := OpenBench.utils.branch_is_out_of_date(workload):
+        warning = 'Consider Rebasing: Dev (%s) appears behind Base (%s)' % (workload.dev.name, workload.base.name)
+
+    username = request.user.username
+    profile  = Profile.objects.get(user=request.user)
+    summary  = 'CREATE P=%d TP=%d' % (workload.priority, workload.throughput)
+    LogEvent.objects.create(author=username, summary=summary, log_file='', test_id=workload.id)
+
+    if not OpenBench.config.USE_CROSS_APPROVAL and profile.approver:
+        workload.approved = True; workload.save()
+
+    return OpenBench.views.redirect(request, '/index/', warning=warning)
+
+
+def create_new_tune(request):
+    return None, ['No endpoint']
+
+def create_new_test(request):
+
+
+    # Collects erros, and collects all data from the Github API
+    errors, engine_info = OpenBench.verify_workload.verify_workload(request, 'TEST')
+    dev_info, dev_has_all = engine_info[0]
+    base_ingo, base_has_all = engine_info[1]
+
+    if errors:
+        return None, errors
+
+    test                   = Test()
+    test.author            = request.user.username
+    test.book_name         = request.POST['book_name']
+
+    test.dev               = get_engine(*dev_info)
+    test.dev_repo          = request.POST['dev_repo']
+    test.dev_engine        = request.POST['dev_engine']
+    test.dev_options       = request.POST['dev_options']
+    test.dev_network       = request.POST['dev_network']
+    test.dev_time_control  = OpenBench.utils.TimeControl.parse(request.POST['dev_time_control'])
+
+    test.base              = get_engine(*base_ingo)
+    test.base_repo         = request.POST['base_repo']
+    test.base_engine       = request.POST['base_engine']
+    test.base_options      = request.POST['base_options']
+    test.base_network      = request.POST['base_network']
+    test.base_time_control = OpenBench.utils.TimeControl.parse(request.POST['base_time_control'])
+
+    test.worker_limit      = 0 if request.POST['worker_limit'] == 'None' else int(request.POST['worker_limit'])
+    test.thread_limit      = 0 if request.POST['thread_limit'] == 'None' else int(request.POST['thread_limit'])
+    test.workload_size     = int(request.POST['workload_size'])
+    test.priority          = int(request.POST['priority'])
+    test.throughput        = int(request.POST['throughput'])
+
+    test.syzygy_wdl        = request.POST['syzygy_wdl']
+    test.syzygy_adj        = request.POST['syzygy_adj']
+    test.win_adj           = request.POST['win_adj']
+    test.draw_adj          = request.POST['draw_adj']
+
+    test.test_mode         = request.POST['test_mode']
+    test.awaiting          = not (dev_has_all and base_has_all)
+
+    if test.test_mode == 'SPRT':
+        test.elolower = float(request.POST['test_bounds'].split(',')[0].lstrip('['))
+        test.eloupper = float(request.POST['test_bounds'].split(',')[1].rstrip(']'))
+        test.alpha    = float(request.POST['test_confidence'].split(',')[1].rstrip(']'))
+        test.beta     = float(request.POST['test_confidence'].split(',')[0].lstrip('['))
+        test.lowerllr = math.log(test.beta / (1.0 - test.alpha))
+        test.upperllr = math.log((1.0 - test.beta) / test.alpha)
+
+    if test.test_mode == 'GAMES':
+        test.max_games = int(request.POST['test_max_games'])
+
+    if test.dev_network:
+        test.dev_netname = Network.objects.get(engine=test.dev_engine, sha256=test.dev_network).name
+
+    if test.base_network:
+        test.base_netname = Network.objects.get(engine=test.base_engine, sha256=test.base_network).name
+
+    test.save()
+
+    profile = Profile.objects.get(user=request.user)
+    profile.tests += 1
+    profile.save()
+
+    return test, None
+
+
+def get_engine(source, name, sha, bench):
+
+    engine = Engine.objects.filter(name=name, source=source, sha=sha, bench=bench)
+    if engine.first() != None:
+        return engine.first()
+
+    return Engine.objects.create(name=name, source=source, sha=sha, bench=bench)
