+++ OpenBench/utils.py
@@ -42,6 +42,8 @@ from OpenBench.config import OPENBENCH_CONFIG
 from OpenBench.models import *
 from OpenBench.stats import TrinomialSPRT, PentanomialSPRT
 
+import OpenBench.verify_workload
+
 import OpenBench.views
 
 
@@ -122,6 +124,13 @@ class TimeControl(object):
         return float(time_str.split('+')[0])
 
 
+
+def read_git_credentials(engine):
+    fname = 'credentials.%s' % (engine.replace(' ', '').lower())
+    if os.path.exists(fname):
+        with open(fname) as fin:
+            return { 'Authorization' : 'token %s' % fin.readlines()[0].rstrip() }
+
 def path_join(*args):
     return "/".join([f.lstrip("/").rstrip("/") for f in args]).rstrip('/')
 
@@ -135,6 +144,8 @@ def extract_option(options, option):
 
     match = re.search('(?<={0}=)[^ ]*'.format(option), options)
     if match: return match.group()
+
+
 
 
 def get_pending_tests():
@@ -209,13 +220,6 @@ def getPaging(content, page, url, pagelen=25):
     return start, end, context
 
 
-def getEngine(source, name, sha, bench):
-
-    engine = Engine.objects.filter(name=name, source=source, sha=sha, bench=bench)
-    if engine.first() != None: return engine.first()
-    return Engine.objects.create(name=name, source=source, sha=sha, bench=bench)
-
-
 
 def branch_is_out_of_date(test):
 
@@ -240,342 +244,6 @@ def branch_is_out_of_date(test):
         traceback.print_exc()
         return False
 
-def read_git_credentials(engine):
-    fname = 'credentials.%s' % (engine.replace(' ', '').lower())
-    if os.path.exists(fname):
-        with open(fname) as fin:
-            return { 'Authorization' : 'token %s' % fin.readlines()[0].rstrip() }
-
-def requests_illegal_fork(request, field):
-
-    # Strip trailing '/'s for sanity
-    engine  = OPENBENCH_CONFIG['engines'][request.POST['%s_engine' % (field)]]
-    eng_src = engine['source'].rstrip('/')
-    tar_src = request.POST['%s_repo' % (field)].rstrip('/')
-
-    # Illegal if sources do not match for Private engines
-    return engine['private'] and eng_src != tar_src
-
-def determine_bench(request, field, message):
-
-    # Use the provided bench if possible
-    try: return int(request.POST['{0}_bench'.format(field)])
-    except: pass
-
-    # Fallback to try to parse the Bench from the commit
-    try:
-        benches = re.findall('(?:BENCH|NODES)[ :=]+([0-9,]+)', message, re.IGNORECASE)
-        return int(benches[-1].replace(',', ''))
-    except: return None
-
-def collect_github_info(request, errors, field):
-
-    # Get branch name / commit sha / tag, and the API path for it
-    branch = request.POST['{0}_branch'.format(field)]
-    bysha  = bool(re.search('[0-9a-fA-F]{40}', branch))
-
-    # All API requests will share this common path. Some engines are private.
-    base    = request.POST['%s_repo' % (field)].replace('github.com', 'api.github.com/repos')
-    engine  = request.POST['%s_engine' % (field)]
-    private = OPENBENCH_CONFIG['engines'][engine]['private']
-    headers = {}
-
-    ## Step 1: Verify the target of the API requests
-    ## [A] We will not attempt to reach any site other than api.github.com
-    ## [B] Private engines may only use their main repo for sources of tests
-    ## [C] Determine which, if any, credentials we want to pass along
-
-    # Private engines must have a token stored in credentials.enginename
-    if private and not (headers := read_git_credentials(engine)):
-        errors.append('Server does not have access tokens for this engine')
-        return (None, None)
-
-    # Do not allow private engines to use forked repos ( We don't have a token! )
-    if requests_illegal_fork(request, field):
-        errors.append('Forked Repositories are not allowed for Private engines')
-        return (None, None)
-
-    # Avoid leaking our credentials to other websites
-    if not base.startswith('https://api.github.com/'):
-        errors.append('OpenBench may only reach Github\'s API')
-        return (None, None)
-
-    ## Step 2: Connect to the Github API for the given Branch or Commit SHA.
-    ## [A] We will attempt to parse the most recent commit message for a
-    ##     bench, unless one was supplied.
-    ## [B] We will translate any branch name into a commit SHA for later use,
-    ##     so we may compare branches and generate diff URLs
-    ## [C] If the engine is public, we will construct the URL to download the
-    ##     source code from Github into a .zip file.
-    ## [D] If the engine is private, we will carry onto Step 3.
-
-    try: # Fetch data from the Github API
-
-        # Lookup branch or commit sha, but will fail for tags
-        url  = path_join(base, 'commits' if bysha else 'branches', branch)
-        data = requests.get(url, headers=headers).json()
-
-        # Check to see if the branch name was actually a tag name
-        if not bysha and 'commit' not in data:
-            url  = path_join(base, 'commits', branch)
-            data = requests.get(url, headers=headers).json()
-
-        # Actual branches have to go one layer deeper
-        elif not bysha: data = data['commit']
-
-        # Check that all the data we need going forward is present
-        assert 'message' in data['commit'] and 'sha' in data
-        assert private or 'sha' in data['commit']['tree']
-
-    except: # Unable to find for whatever reason
-        import traceback
-        traceback.print_exc()
-        errors.append('%s could not be found' % (branch))
-        return (None, None)
-
-    # Extract the bench from the web form, or from the commit message
-    if not (bench := determine_bench(request, field, data['commit']['message'])):
-        errors.append('Unable to parse a Bench for %s' % (branch))
-        return (None, None)
-
-    # Public Engines: Construct the .zip download and return everything
-    if not private:
-        treeurl = data['commit']['tree']['sha'] + '.zip'
-        source  = path_join(request.POST['%s_repo' % (field)], 'archive', treeurl)
-        return (source, branch, data['sha'], bench), True
-
-    ## Step 3: Construct the URL for the API request to list all Artifacts
-    ## [A] OpenBench artifacts are always run via a file named openbench.yml
-    ## [B] These should contain combinations for windows/linux, avx2/avx512, popcnt/pext
-    ## [C] If those artifacts are not found, we flag the test as awaiting, and try later.
-
-    url, has_all = fetch_artifact_url(base, engine, headers, data['sha'])
-    return (url, branch, data['sha'], bench), has_all
-
-def fetch_artifact_url(base, engine, headers, sha):
-
-    try:
-        # Fetch the run id for the openbench workflow for this comment
-        url    = path_join(base, 'actions', 'workflows', 'openbench.yml', 'runs')
-        url   += '?head_sha=%s' % (sha)
-        run_id = requests.get(url=url, headers=headers).json()['workflow_runs'][0]['id']
-
-        # Fetch information about individual job results
-        url  = path_join(base, 'actions', 'runs', str(run_id), 'jobs')
-        jobs = requests.get(url=url, headers=headers).json()['jobs']
-
-        # Fetch information about individual artifact
-        url       = path_join(base, 'actions', 'runs', str(run_id), 'artifacts')
-        artifacts = requests.get(url=url, headers=headers).json()['artifacts']
-
-        # All jobs finished, with at least one non-expired Artifact
-        assert not any(job['conclusion'] != 'success' for job in jobs)
-        assert not any(artifact['expired'] for artifact in artifacts)
-        assert len(artifacts) >= len(jobs)
-
-        # Only set the url if we have everything we need
-        return (url, True)
-
-    except Exception as error:
-        # If anything goes wrong, retry later with the same base URL
-        return (base, False)
-
-def verify_test_creation(request):
-
-    errors = []
-
-    def verify_integer(field, field_name):
-        try: int(request.POST[field])
-        except: errors.append('"{0}" is not an Integer'.format(field_name))
-
-    def verify_integer_or_none(field, field_name):
-        try: True if request.POST[field] == 'None' else int(request.POST[field])
-        except: errors.append('"{0}" is neither an Integer nor "None"'.format(field_name))
-
-    def verify_greater_than(field, field_name, value):
-        try: assert int(request.POST[field]) > value
-        except: errors.append('"{0}" is not greater than {1}'.format(field_name, value))
-
-    def verify_options(field, option, field_name):
-        try: assert int(extract_option(request.POST[field], option)) >= 1
-        except: errors.append('"{0}" needs to be at least 1 for {1}'.format(option, field_name))
-
-    def verify_configuration(field, field_name, parent):
-        try: assert request.POST[field] in OPENBENCH_CONFIG[parent].keys()
-        except: errors.append('{0} was not found in the configuration'.format(field_name))
-
-    def verify_time_control(field, field_name):
-        try: TimeControl.parse(request.POST[field])
-        except: errors.append('{0} is not parsable'.format(field_name))
-
-    def verify_win_adj(field):
-        try:
-            if (content := request.POST[field]) == 'None': return
-            assert re.match('movecount=[0-9]+ score=[0-9]+', content)
-        except: errors.append('Invalid Win Adjudication Setting. Try "None"?')
-
-    def verify_draw_adj(field):
-        try:
-            if (content := request.POST[field]) == 'None': return
-            assert re.match('movenumber=[0-9]+ movecount=[0-9]+ score=[0-9]+', content)
-        except: errors.append('Invalid Draw Adjudication Setting. Try "None"?')
-
-    def verify_github_repo(field):
-        pattern = r'^https:\/\/github\.com\/[A-Za-z0-9-]+\/[A-Za-z0-9_.-]+\/?$'
-        try: assert re.match(pattern, request.POST[field])
-        except: errors.append('Sources must be found on https://github.com/<User>/<Repo>')
-
-    def verify_network(field, field_name, engine_field):
-        try:
-            if request.POST[field] == '': return
-            Network.objects.get(engine=request.POST[engine_field], sha256=request.POST[field])
-        except: errors.append('Unknown Network Provided for {0}'.format(field_name))
-
-    def verify_test_mode(field):
-        try: assert request.POST[field] in ['SPRT', 'GAMES']
-        except: errors.append('Unknown Test Mode')
-
-    def verify_sprt_bounds(field):
-        try:
-            if request.POST['test_mode'] != 'SPRT': return
-            pattern = r'^\[(-?\d+(?:\.\d+)?), (-?\d+(?:\.\d+)?)\]$'
-            match   = re.match(pattern, request.POST['test_bounds'])
-            assert float(match.group(1)) < float(match.group(2))
-        except: errors.append('SPRT Bounds must be formatted as [float1, float2]')
-
-    def verify_sprt_conf(field):
-        try:
-            if request.POST['test_mode'] != 'SPRT': return
-            pattern = r'^\[(-?\d+(?:\.\d+)?), (-?\d+(?:\.\d+)?)\]$'
-            match   = re.match(pattern, request.POST['test_confidence'])
-            assert 0.00 < float(match.group(1)) < 1.00
-            assert 0.00 < float(match.group(2)) < 1.00
-        except: errors.append('Confidence Bounds must be formatted as [float1, float2], within (0.00, 1.00)')
-
-    def verify_max_games(field):
-        try:
-            if request.POST['test_mode'] != 'GAMES': return
-            assert int(request.POST['test_max_games']) > 0
-        except: errors.append('Fixed Games Tests must last at least one game')
-
-    def verify_syzygy_field(field, field_name):
-        candidates = ['OPTIONAL', 'DISABLED', '3-MAN', '4-MAN', '5-MAN', '6-MAN', '7-MAN']
-        try: assert request.POST[field] in candidates
-        except: errors.append('%s must be in %s' % (field_name, ', '.join(candidates)))
-
-    verifications = [
-
-        # Verify everything about the Dev Engine
-        (verify_configuration  , 'dev_engine', 'Dev Engine', 'engines'),
-        (verify_github_repo    , 'dev_repo'),
-        (verify_network        , 'dev_network', 'Dev Network', 'dev_engine'),
-        (verify_options        , 'dev_options', 'Threads', 'Dev Options'),
-        (verify_options        , 'dev_options', 'Hash', 'Dev Options'),
-        (verify_time_control   , 'dev_time_control', 'Dev Time Control'),
-
-        # Verify everything about the Base Engine
-        (verify_configuration  , 'base_engine', 'Base Engine', 'engines'),
-        (verify_github_repo    , 'base_repo'),
-        (verify_network        , 'base_network', 'Base Network', 'base_engine'),
-        (verify_options        , 'base_options', 'Threads', 'Base Options'),
-        (verify_options        , 'base_options', 'Hash', 'Base Options'),
-        (verify_time_control   , 'base_time_control', 'Base Time Control'),
-
-        # Verify everything about the Test Settings
-        (verify_configuration  , 'book_name', 'Book', 'books'),
-        (verify_test_mode      , 'test_mode'),
-        (verify_sprt_bounds    , 'test_bounds'),
-        (verify_sprt_conf      , 'test_confidence'),
-        (verify_max_games      , 'test_max_games'),
-
-        # Verify everything about the General Settings
-        (verify_integer        , 'priority', 'Priority'),
-        (verify_greater_than   , 'throughput', 'Throughput', 0),
-        (verify_syzygy_field   , 'syzygy_wdl', 'Syzygy WDL'),
-
-        # Verify everything about the Workload Settings
-        (verify_integer_or_none, 'worker_limit', 'Worker Limit'),
-        (verify_integer_or_none, 'thread_limit', 'Thread Limit'),
-        (verify_greater_than   , 'workload_size', 'Workload Size', 0),
-
-        # Verify everything about the Adjudicaton Settings
-        (verify_syzygy_field   , 'syzygy_adj', 'Syzygy Adjudication'),
-        (verify_win_adj        , 'win_adj'),
-        (verify_draw_adj       , 'draw_adj'),
-    ]
-
-    for verification in verifications:
-        verification[0](*verification[1:])
-
-    return errors
-
-def create_new_test(request):
-
-    # Verify all of the fields in the request
-    if (errors := verify_test_creation(request)):
-        return None, errors
-
-    # Collect Github meta information (Shas, Sources, Benches)
-    devinfo,  dev_has_all  = collect_github_info(request, errors, 'dev')
-    baseinfo, base_has_all = collect_github_info(request, errors, 'base')
-    if errors != []: return None, errors
-
-    test                   = Test()
-    test.author            = request.user.username
-    test.book_name         = request.POST['book_name']
-
-    test.dev               = getEngine(*devinfo)
-    test.dev_repo          = request.POST['dev_repo']
-    test.dev_engine        = request.POST['dev_engine']
-    test.dev_options       = request.POST['dev_options']
-    test.dev_network       = request.POST['dev_network']
-    test.dev_time_control  = TimeControl.parse(request.POST['dev_time_control'])
-
-    test.base              = getEngine(*baseinfo)
-    test.base_repo         = request.POST['base_repo']
-    test.base_engine       = request.POST['base_engine']
-    test.base_options      = request.POST['base_options']
-    test.base_network      = request.POST['base_network']
-    test.base_time_control = TimeControl.parse(request.POST['base_time_control'])
-
-    test.worker_limit      = 0 if request.POST['worker_limit'] == 'None' else int(request.POST['worker_limit'])
-    test.thread_limit      = 0 if request.POST['thread_limit'] == 'None' else int(request.POST['thread_limit'])
-    test.workload_size     = int(request.POST['workload_size'])
-    test.priority          = int(request.POST['priority'])
-    test.throughput        = int(request.POST['throughput'])
-
-    test.syzygy_wdl        = request.POST['syzygy_wdl']
-    test.syzygy_adj        = request.POST['syzygy_adj']
-    test.win_adj           = request.POST['win_adj']
-    test.draw_adj          = request.POST['draw_adj']
-
-    test.test_mode         = request.POST['test_mode']
-    test.awaiting          = not (dev_has_all and base_has_all)
-
-    if test.test_mode == 'SPRT':
-        test.elolower = float(request.POST['test_bounds'].split(',')[0].lstrip('['))
-        test.eloupper = float(request.POST['test_bounds'].split(',')[1].rstrip(']'))
-        test.alpha    = float(request.POST['test_confidence'].split(',')[1].rstrip(']'))
-        test.beta     = float(request.POST['test_confidence'].split(',')[0].lstrip('['))
-        test.lowerllr = math.log(test.beta / (1.0 - test.alpha))
-        test.upperllr = math.log((1.0 - test.beta) / test.alpha)
-
-    if test.test_mode == 'GAMES':
-        test.max_games = int(request.POST['test_max_games'])
-
-    if test.dev_network:
-        test.dev_netname = Network.objects.get(engine=test.dev_engine, sha256=test.dev_network).name
-
-    if test.base_network:
-        test.base_netname = Network.objects.get(engine=test.base_engine, sha256=test.base_network).name
-
-    test.save()
-
-    profile = Profile.objects.get(user=request.user)
-    profile.tests += 1
-    profile.save()
-
-    return test, None
 
 
 def get_machine(machineid, user, info):
