+++ OpenBench/watcher.py
@@ -25,6 +25,7 @@ import time
 import traceback
 
 import OpenBench.utils
+import OpenBench.verify_workload
 
 class ArtifactWatcher(threading.Thread):
 
@@ -42,13 +43,13 @@ class ArtifactWatcher(threading.Thread):
         if not dev_has: # Check for new Artifacts for Dev
             dev_headers = OpenBench.utils.read_git_credentials(test.dev_engine)
             data = [test.dev.source, test.dev_engine, dev_headers, test.dev.sha]
-            test.dev.source, dev_has = OpenBench.utils.fetch_artifact_url(*data)
+            test.dev.source, dev_has = OpenBench.verify_workload.fetch_artifact_url(*data)
             test.dev.save()
 
         if not base_has: # Check for new Artifacts for Base
             base_headers = OpenBench.utils.read_git_credentials(test.base_engine)
             data = [test.base.source, test.base_engine, base_headers, test.base.sha]
-            test.base.source, base_has = OpenBench.utils.fetch_artifact_url(*data)
+            test.base.source, base_has = OpenBench.verify_workload.fetch_artifact_url(*data)
             test.base.save()
 
         # If both finished, flag the test as no longer awaiting
