+++ Templates/OpenBench/create_tune.html
@@ -22,7 +22,7 @@
 
 {% block content %}
 
-    <form method="POST" action="/newTest/">
+    <form method="POST" action="/newTune/">
         {% csrf_token %}
 
         <div class="form w-100">
